chromium/third_party/blink/web_tests/external/wpt/html/browsers/windows/nested-browsing-contexts/frameElement-siblings.sub-expected.txt

This is a testharness.js-based test.
PASS it must give a "SecurityError" DOMException if the pages are different-origin domain
FAIL it must return the iframe element if the pages are same-origin domain assert_equals: expected "HTMLIFrameElement" but got "SecurityError"
Harness: the test ran to completion.
